+++ lance-spark-base_2.12/src/main/java/org/lance/spark/internal/LanceFragmentScanner.java
@@ -17,6 +17,7 @@ import org.lance.Fragment;
 import org.lance.ipc.LanceScanner;
 import org.lance.ipc.ScanOptions;
 import org.lance.spark.LanceConstant;
+import org.lance.spark.LanceRuntime;
 import org.lance.spark.LanceSparkReadOptions;
 import org.lance.spark.read.LanceInputPartition;
 
@@ -50,7 +51,7 @@ public class LanceFragmentScanner implements AutoCloseable {
     try {
       LanceSparkReadOptions readOptions = inputPartition.getReadOptions();
       Fragment fragment =
-          LanceDatasetCache.getFragment(
+          LanceRuntime.getFragment(
               readOptions,
               fragmentId,
               inputPartition.getInitialStorageOptions(),
